+++ samples/Kovecses.Requests.Sample/Features/Books/GetBooks/GetBooksHandler.cs
@@ -2,6 +2,6 @@ namespace Kovecses.Requests.Sample.Features.Books.GetBooks;
 
 internal sealed class GetBooksHandler(IBookRepository repository) : IRequestHandler<GetBooksQuery, IEnumerable<BookDto>>
 {
-    public Task<IEnumerable<BookDto>> Handle(GetBooksQuery request, CancellationToken cancellationToken)
+    public Task<IEnumerable<BookDto>> HandleAsync(GetBooksQuery request, CancellationToken cancellationToken)
         => repository.GetAllAsync(cancellationToken);
 }